redux関連の修正
